Remove circular dependency between `json` and `util`
Extracting `object-util` resolve this circular dependency issue.
