+++ packages/core/src/internal/util.ts
@@ -24,10 +24,6 @@ import { stringify } from '../json'
 
 const ENCRYPTION_ON: EncryptionLevel = 'ENCRYPTION_ON'
 const ENCRYPTION_OFF: EncryptionLevel = 'ENCRYPTION_OFF'
-// eslint-disable-next-line @typescript-eslint/naming-convention
-const __isBrokenObject__ = '__isBrokenObject__'
-// eslint-disable-next-line @typescript-eslint/naming-convention
-const __reason__ = '__reason__'
 /**
  * Verifies if the object is null or empty
  * @param obj The subject object
@@ -226,65 +222,6 @@ function isString(str: any): str is string {
   return Object.prototype.toString.call(str) === '[object String]'
 }
 
-/**
- * Creates a object which all method call will throw the given error
- * 
- * @param {Error} error The error
- * @param {any} object The object. Default: {}
- * @returns {any} A broken object
- */
-function createBrokenObject<T extends object> (error: Error, object: any = {}): T {
-  const fail = () => {
-    throw error
-  }
-
-  return new Proxy(object, {
-    get: (_: T, p: string | Symbol): any => {
-      if (p === __isBrokenObject__) {
-        return true
-      } else if (p === __reason__) {
-        return error
-      } else if (p === 'toJSON') {
-        return undefined
-      }
-      fail()
-    },
-    set: fail,
-    apply: fail,
-    construct: fail,
-    defineProperty: fail,
-    deleteProperty: fail,
-    getOwnPropertyDescriptor: fail,
-    getPrototypeOf: fail,
-    has: fail,
-    isExtensible: fail,
-    ownKeys: fail,
-    preventExtensions: fail,
-    setPrototypeOf: fail,
-  })
-}
-
-/**
- * Verifies if it is a Broken Object
- * @param {any} object The object
- * @returns {boolean} If it was created with createBrokenObject
- */
-function isBrokenObject (object: any): boolean {
-  return object !== null && typeof object === 'object' && object[__isBrokenObject__] === true
-}
-
-/**
- * Returns if the reason the object is broken.
- *
- * This method should only be called with instances create with {@link createBrokenObject}
- *
- * @param {any} object The object
- * @returns {Error} The reason the object is broken
- */
-function getBrokenObjectReason (object: any): Error {
-  return object[__reason__]
-}
-
 export {
   isEmptyObjectOrNull,
   isObject,
@@ -296,8 +233,5 @@ export {
   assertValidDate,
   validateQueryAndParameters,
   ENCRYPTION_ON,
-  ENCRYPTION_OFF,
-  createBrokenObject,
-  isBrokenObject,
-  getBrokenObjectReason
+  ENCRYPTION_OFF
 }
